MVS extension: renamed `additions` to `components`
