+++ src/extensions/mvs/behavior.ts
@@ -13,13 +13,13 @@ import { PluginContext } from '../../mol-plugin/context';
 import { StructureRepresentationProvider } from '../../mol-repr/structure/representation';
 import { ColorTheme } from '../../mol-theme/color';
 import { ParamDefinition as PD } from '../../mol-util/param-definition';
-import { MVSAnnotationColorThemeProvider } from './additions/annotation-color-theme';
-import { MVSAnnotationLabelRepresentationProvider } from './additions/annotation-label/representation';
-import { MVSAnnotationsProvider } from './additions/annotation-prop';
-import { MVSAnnotationTooltipsLabelProvider, MVSAnnotationTooltipsProvider } from './additions/annotation-tooltips-prop';
-import { CustomLabelRepresentationProvider } from './additions/custom-label/representation';
-import { CustomTooltipsLabelProvider, CustomTooltipsProvider } from './additions/custom-tooltips-prop';
-import { makeMultilayerColorThemeProvider } from './additions/multilayer-color-theme';
+import { MVSAnnotationColorThemeProvider } from './components/annotation-color-theme';
+import { MVSAnnotationLabelRepresentationProvider } from './components/annotation-label/representation';
+import { MVSAnnotationsProvider } from './components/annotation-prop';
+import { MVSAnnotationTooltipsLabelProvider, MVSAnnotationTooltipsProvider } from './components/annotation-tooltips-prop';
+import { CustomLabelRepresentationProvider } from './components/custom-label/representation';
+import { CustomTooltipsLabelProvider, CustomTooltipsProvider } from './components/custom-tooltips-prop';
+import { makeMultilayerColorThemeProvider } from './components/multilayer-color-theme';
 import { loadMVS } from './load';
 import { MVSData } from './mvs-data';
