+++ src/extensions/mvs/components/annotation-label/representation.ts
