+++ src/extensions/mvs/components/annotation-label/visual.ts
