+++ src/extensions/mvs/components/annotation-structure-component.ts
