+++ src/extensions/mvs/components/annotation-tooltips-prop.ts
