+++ src/extensions/mvs/components/custom-label/visual.ts
