+++ src/extensions/mvs/load-helpers.ts
@@ -12,14 +12,14 @@ import { PluginContext } from '../../mol-plugin/context';
 import { StateBuilder, StateObjectSelector, StateTransformer } from '../../mol-state';
 import { arrayDistinct } from '../../mol-util/array';
 import { canonicalJsonString } from '../../mol-util/object';
-import { MVSAnnotationColorThemeProps, MVSAnnotationColorThemeProvider } from './additions/annotation-color-theme';
-import { MVSAnnotationLabelRepresentationProvider } from './additions/annotation-label/representation';
-import { MVSAnnotationSpec } from './additions/annotation-prop';
-import { MVSAnnotationStructureComponentProps } from './additions/annotation-structure-component';
-import { MVSAnnotationTooltipsProps } from './additions/annotation-tooltips-prop';
-import { CustomTooltipsProps } from './additions/custom-tooltips-prop';
-import { MultilayerColorThemeName, MultilayerColorThemeProps, NoColor } from './additions/multilayer-color-theme';
-import { SelectorAll } from './additions/selector';
+import { MVSAnnotationColorThemeProps, MVSAnnotationColorThemeProvider } from './components/annotation-color-theme';
+import { MVSAnnotationLabelRepresentationProvider } from './components/annotation-label/representation';
+import { MVSAnnotationSpec } from './components/annotation-prop';
+import { MVSAnnotationStructureComponentProps } from './components/annotation-structure-component';
+import { MVSAnnotationTooltipsProps } from './components/annotation-tooltips-prop';
+import { CustomTooltipsProps } from './components/custom-tooltips-prop';
+import { MultilayerColorThemeName, MultilayerColorThemeProps, NoColor } from './components/multilayer-color-theme';
+import { SelectorAll } from './components/selector';
 import { rowToExpression, rowsToExpression } from './helpers/selections';
 import { ElementOfSet, decodeColor, isDefined, stringHash } from './helpers/utils';
 import { MolstarLoadingContext } from './load';
